+++ NanoChallenge3/Model/SavingAccount.swift
@@ -11,8 +11,8 @@ struct SavingAccount {
     var documentID: String?
     var name: String = ""
     var hue: Float = 0.0
-    var goal: Double = 0.0
-    var balance: Double = 0.0
+    var goal: Int = 0
+    var balance: Int = 0
 }
 
 extension SavingAccount: DocumentSerializable {
@@ -20,8 +20,8 @@ extension SavingAccount: DocumentSerializable {
     init?(documentID: String, data: [String : Any]) {
         guard let name = data["name"] as? String,
             let hue = data["hue"] as? Float,
-            let goal = data["goal"] as? Double,
-            let balance = data["balance"] as? Double else { return nil }
+            let goal = data["goal"] as? Int,
+            let balance = data["balance"] as? Int else { return nil }
         
         self.init(documentID: documentID, name: name, hue: hue, goal: goal, balance: balance)
     }
