+++ NanoChallenge3/Modules/Investments/ViewModel/InstitutionListViewModel.swift
@@ -30,7 +30,7 @@ final class InstitutionListViewModel: ViewModel {
         return institutions[index].hue
     }
     
-    func balance(forInstitutionAtIndex index: Int) -> Double {
+    func balance(forInstitutionAtIndex index: Int) -> Int {
         return institutions[index].balance
     }
